use abstract class for observer
