+++ src/search/observers/observers.hpp
@@ -1,0 +1,18 @@
+#ifndef STASE_OBSERVERS_HPP
+#define STASE_OBSERVERS_HPP
+
+#include "search.h"
+
+/**
+ *
+ * This header collects together various different observer implementations: they
+ * are each defined nicely in their own files, and get drawn together here.
+ *
+ */
+
+#include "null_observer.hpp"
+
+#include "xml_observer.hpp"
+
+
+#endif //STASE_OBSERVERS_HPP
