+++ src/search/observers/xml_observer.hpp
@@ -1,57 +1,33 @@
+#ifndef STASE_XML_OBSERVER_HPP
+#define STASE_XML_OBSERVER_HPP
+
 #include <fstream>
 #include "search.h"
-#include "../utils/utils.h"
-
-enum SearchEvent {
-    VISIT,
-    VISIT_LINE,
-    FORCE_VISIT,
-    FORCE_VISIT_LINE,
-    DEEPEN
-};
-
-inline std::string name(SearchEvent ev) {
-    switch (ev) {
-        case VISIT: return "visit";
-        case VISIT_LINE: return "visit_l";
-        case FORCE_VISIT: return "f_visit";
-        case FORCE_VISIT_LINE: return "f_visit_l";
-        case DEEPEN: return "deepen";
-        default: return "unknown event";
-    }
-}
-
-class NullObserver {
-public:
-    inline void open_event(const SearchNode * node, SearchEvent ev) {}
-    inline void close_event(const SearchNode * node, SearchEvent ev) {}
-};
+#include "../../utils/utils.h"
 
 
-class Observer {
+class XMLObserver : public Observer {
 
 private:
     const std::string filepath;
     std::vector<std::string> current_line;
     std::vector<std::string> buffer;
     int indent_level = 1;
-    const SearchNode * parent = nullptr;
 
 public:
 
-    Observer(const std::string & filepath) : filepath(filepath) {}
+    XMLObserver(const std::string & filepath) : filepath(filepath) {}
 
-    inline void open_event(const SearchNode * node, SearchEvent ev) {
+    inline void open_event(const SearchNode * node, const SearchEvent ev) {
         const std::string move = (is_sentinel(node->move))
                 ? ""
                 : '"' + move2uci(node->move) + '"';
         current_line.push_back(move2uci(node->move));
         buffer.push_back(indent(indent_level) + "<" + name(ev) + ">" + move);
         ++indent_level;
-        parent = node;
     }
 
-    inline void close_event(const SearchNode * node, SearchEvent ev) {
+    inline void close_event(const SearchNode * node, const SearchEvent ev) {
         const std::string move = (is_sentinel(node->move))
                  ? ""
                  : '"' + move2uci(node->move) + '"';
@@ -97,3 +73,5 @@ private:
         return s;
     }
 };
+
+#endif // STASE_XML_OBSERVER_HPP
