+++ src/test/search/test_observer.hpp
@@ -1,0 +1,14 @@
+#ifndef STASE_TEST_OBSERVER_HPP
+#define STASE_TEST_OBSERVER_HPP
+
+#include "../../search/observers/observers.hpp"
+
+class CandListOrderingObserver : Observer {
+
+    inline void open_event(const SearchNode * node, SearchEvent ev) {
+
+    }
+
+};
+
+#endif //STASE_TEST_OBSERVER_HPP
